Refactor code in app file
