+++ src/app/assets/classes.py
